Browser hangs after selecting a target in Shortcut and Relationships content #616
